Add files via upload
front_obstacle_distance: This variable stores the distance to the obstacle as measured by the front lidar.

Print Statement: It prints the obstacle distance from the front lidar to the console.

Time Calculation (current_time and det_t): current_time is the timestamp of the current iteration. det_t is the time difference between the current and previous iterations. This time difference is used in the PID controller.

Condition Check: It checks if the front_obstacle_distance is less than the threshold (self.front_lidar_threshold). If true, it means there is an obstacle detected by the front lidar.

Height Adjustment (des): The desired height (des) is calculated as the sum of the barometric height (baro_height) and the difference between the maintain_threshold and the front_obstacle_distance. If des is less than the barro_threshold, it is set to barro_threshold.

PID Controller (pwm): The PID controller (self.pid) is then used to compute the PWM signal (pwm) based on the barometric height (baro_height) and the desired height (des). The time difference (det_t) is also passed to the PID controller.

Position Update (self.pos_update): The PWM signal is published to control the drone's altitude, and the current barometric height is stored.